Django/django-school-api-VI/school_proj/requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.0
psycopg==3.1.18
psycopg-binary==3.1.18
sqlparse==0.4.4
typing_extensions==4.10.0
